app/requirements.txt
Flask==1.1.2
python-dotenv==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
PyMySQL==0.10.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
Flask-Admin==1.5.6
flask-ldap3-login==0.9.16
email-validator==1.1.1
Flask-Principal==0.4.0
Flask-DebugToolbar==0.11.0
Flask-Caching==1.9.0
Flask-Static-Digest==0.1.4
webpreview==1.6.0
redis==3.5.3
cryptography==3.2.1
Pillow==8.0.1
IPTCInfo3==2.1.4
apifairy==0.6.2
marshmallow-sqlalchemy==0.24.1
